PB-114: Move featureStyleFunction out of drawing module
This function is used in several modules/utils
