+++ src/modules/drawing/lib/export-utils.js
@@ -4,9 +4,9 @@ import { LineString, Polygon } from 'ol/geom'
 import { Circle, Icon } from 'ol/style'
 import Style from 'ol/style/Style'
 
-import { featureStyleFunction } from '@/modules/drawing/lib/style'
 import i18n from '@/modules/i18n/index'
 import { WGS84 } from '@/utils/coordinates/coordinateSystems'
+import { featureStyleFunction } from '@/utils/featureStyleUtils'
 import log from '@/utils/logging'
 
 const kmlFormat = new KML()
